+++ app/routers/emblem_delete_router.py
@@ -4,8 +4,8 @@ from app.database import get_session
 from app.cache import get_cache
 from app.decorators.locked_decorator import locked
 from app.models.user_model import User, UserRole
-from app.models.member_model import Member
-from app.schemas.member_schemas import EmblemDeleteResponse
+from app.models.partner_model import Partner
+from app.schemas.partner_schemas import EmblemDeleteResponse
 from app.errors import E
 from app.hooks import Hook
 from app.auth import auth
@@ -18,32 +18,32 @@ from app.constants import (
 router = APIRouter()
 
 
-@router.delete("/member/{member_id}/emblem", summary="Remove member emblem",
+@router.delete("/partner/{partner_id}/emblem", summary="Remove partner emblem",
                response_class=JSONResponse, status_code=status.HTTP_200_OK,
-               response_model=EmblemDeleteResponse, tags=["Members"])
+               response_model=EmblemDeleteResponse, tags=["Partners"])
 @locked
 async def emblem_delete(
-    member_id: int,
+    partner_id: int,
     session=Depends(get_session), cache=Depends(get_cache),
     current_user: User = Depends(auth(UserRole.editor))
 ) -> EmblemDeleteResponse:
-    member_repository = Repository(session, cache, Member)
-    member = await member_repository.select(id=member_id)
+    partner_repository = Repository(session, cache, Partner)
+    partner = await partner_repository.select(id=partner_id)
 
-    if not member:
-        raise E([LOC_PATH, "member_id"], member_id,
+    if not partner:
+        raise E([LOC_PATH, "partner_id"], partner_id,
                 ERR_RESOURCE_NOT_FOUND, status.HTTP_404_NOT_FOUND)
 
     hook = Hook(session, cache, current_user=current_user)
-    await hook.do(HOOK_BEFORE_EMBLEM_DELETE, member)
+    await hook.do(HOOK_BEFORE_EMBLEM_DELETE, partner)
 
-    if member.emblem_filename:
-        await FileManager.delete(member.emblem_path)
+    if partner.emblem_filename:
+        await FileManager.delete(partner.emblem_path)
 
-    member.emblem_filename = None
-    await member_repository.update(member, commit=False)
+    partner.emblem_filename = None
+    await partner_repository.update(partner, commit=False)
 
-    await member_repository.commit()
-    await hook.do(HOOK_AFTER_EMBLEM_DELETE, member)
+    await partner_repository.commit()
+    await hook.do(HOOK_AFTER_EMBLEM_DELETE, partner)
 
-    return {"member_id": member.id}
+    return {"partner_id": partner.id}