+++ app/routers/partner_select_router.py
@@ -4,35 +4,35 @@ from app.database import get_session
 from app.cache import get_cache
 from app.decorators.locked_decorator import locked
 from app.models.user_model import User, UserRole
-from app.models.member_model import Member
-from app.schemas.member_schemas import MemberSelectResponse
+from app.models.partner_model import Partner
+from app.schemas.partner_schemas import PartnerSelectResponse
 from app.repository import Repository
 from app.errors import E
 from app.hooks import Hook
 from app.auth import auth
 from app.constants import (
-    LOC_PATH, ERR_RESOURCE_NOT_FOUND, HOOK_AFTER_MEMBER_SELECT)
+    LOC_PATH, ERR_RESOURCE_NOT_FOUND, HOOK_AFTER_PARTNER_SELECT)
 
 router = APIRouter()
 
 
-@router.get("/member/{member_id}", summary="Retrieve member",
+@router.get("/partner/{partner_id}", summary="Retrieve partner",
             response_class=JSONResponse, status_code=status.HTTP_200_OK,
-            response_model=MemberSelectResponse, tags=["Members"])
+            response_model=PartnerSelectResponse, tags=["Partners"])
 @locked
-async def member_select(
-    member_id: int,
+async def partner_select(
+    partner_id: int,
     session=Depends(get_session), cache=Depends(get_cache),
     current_user: User = Depends(auth(UserRole.reader)),
-) -> MemberSelectResponse:
-    member_repository = Repository(session, cache, Member)
-    member = await member_repository.select(id=member_id)
+) -> PartnerSelectResponse:
+    partner_repository = Repository(session, cache, Partner)
+    partner = await partner_repository.select(id=partner_id)
 
-    if not member:
-        raise E([LOC_PATH, "member_id"], member_id,
+    if not partner:
+        raise E([LOC_PATH, "partner_id"], partner_id,
                 ERR_RESOURCE_NOT_FOUND, status.HTTP_404_NOT_FOUND)
 
     hook = Hook(session, cache, current_user=current_user)
-    await hook.do(HOOK_AFTER_MEMBER_SELECT, member)
+    await hook.do(HOOK_AFTER_PARTNER_SELECT, partner)
 
-    return member.to_dict()
+    return partner.to_dict()