+++ app/schemas/document_schemas.py
@@ -7,7 +7,7 @@ documents.
 from typing import Optional, Literal, List, Union
 from pydantic import BaseModel, Field, field_validator
 from app.schemas.user_schemas import UserSelectResponse
-from app.schemas.member_schemas import MemberSelectResponse
+from app.schemas.partner_schemas import PartnerSelectResponse
 from app.schemas.collection_schemas import CollectionSelectResponse
 from app.schemas.revision_schemas import RevisionSelectResponse
 from app.validators.document_validators import (
@@ -37,7 +37,7 @@ class DocumentSelectResponse(BaseModel):
     updated_date: int
     user_id: int
     collection_id: Optional[int]
-    member_id: Optional[int]
+    partner_id: Optional[int]
 
     document_name: str
     document_summary: Optional[str] = None
@@ -49,7 +49,7 @@ class DocumentSelectResponse(BaseModel):
     document_tags: list
     document_user: UserSelectResponse
     document_collection: Optional[CollectionSelectResponse] = None
-    document_member: Optional[MemberSelectResponse] = None
+    document_partner: Optional[PartnerSelectResponse] = None
     latest_revision: RevisionSelectResponse
 
 
@@ -60,7 +60,7 @@ class DocumentUpdateRequest(BaseModel):
     optionally the document name, summary, and tags.
     """
     collection_id: Optional[int] = None
-    member_id: Optional[int] = None
+    partner_id: Optional[int] = None
     document_name: str = Field(..., min_length=1, max_length=256)
     document_summary: Optional[str] = Field(max_length=512, default=None)
     tags: Optional[str] = Field(max_length=256, default=None)